temperature display float ok

+++ android-nrf-toolbox-master/app/src/main/java/no/nordicsemi/android/nrftoolbox/template/TemplateManagerCallbacks.java
@@ -48,6 +48,6 @@ public interface TemplateManagerCallbacks extends BleManagerCallbacks {
 	 * @param value
 	 *            the new value
 	 */
-	void onSampleValueReceived(final BluetoothDevice device, int value);
+	void onSampleValueReceived(final BluetoothDevice device, float value);
 
 }
